+++ commands/themelio-client/src/shell/prompt.rs
@@ -1,5 +1,5 @@
-use crate::common::prompt::{common_read_line, InputPrompt};
-use crate::wallet_shell::command::ShellCommand;
+use crate::utils::prompt::{common_read_line, InputPrompt};
+use crate::shell::command::ShellCommand;
 use async_trait::async_trait;
 use colored::Colorize;
 use std::convert::TryFrom;
