+++ commands/themelio-client/src/utils/formatter/json.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeMap;
 
 use blkstructs::{CoinDataHeight, CoinID};
 
-use crate::common::formatter::formatter::OutputFormatter;
+use crate::utils::formatter::formatter::OutputFormatter;
 use crate::wallet::data::WalletData;
 use crate::wallet::wallet::Wallet;
 use async_trait::async_trait;
