+++ commands/themelio-client/src/wallet/wallet.rs
@@ -1,4 +1,4 @@
-use crate::common::context::ExecutionContext;
+use crate::utils::context::ExecutionContext;
 use crate::wallet::data::WalletData;
 use blkstructs::{
     CoinData, CoinDataHeight, CoinID, Transaction, TxKind, DENOM_TMEL, MICRO_CONVERTER,
